honksquad: refactor: split cybernetic organ effects tests into per-organ files
Break up the monolithic CyberneticOrganEffectsTest (1068 lines, 25 tests)
into 6 focused test files, one per organ type: heart, eyes, lungs, stomach,
ears, liver. Each file carries only the prototypes it needs.
